track: move resolve to better merge
