+++ src/main/java/raylras/zen/code/resolve/DeclaratorResolver.java
@@ -1,4 +1,4 @@
-package raylras.zen.code.type.resolve;
+package raylras.zen.code.resolve;
 
 import org.antlr.v4.runtime.tree.ParseTree;
 import raylras.zen.code.symbol.Declarator;
