+++ src/main/java/raylras/zen/code/symbol/FunctionSymbol.java
@@ -5,9 +5,9 @@ import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.tree.ParseTree;
 import raylras.zen.code.CompilationUnit;
 import raylras.zen.code.parser.ZenScriptParser;
-import raylras.zen.code.type.resolve.NameResolver;
-import raylras.zen.code.type.resolve.ParamsResolver;
-import raylras.zen.code.type.resolve.ReturnTypeResolver;
+import raylras.zen.code.resolve.NameResolver;
+import raylras.zen.code.resolve.ParamsResolver;
+import raylras.zen.code.resolve.ReturnTypeResolver;
 import raylras.zen.code.type.FunctionType;
 import raylras.zen.code.type.Type;
 import raylras.zen.util.SymbolUtils;
