+++ src/main/java/raylras/zen/code/symbol/Symbol.java
@@ -2,7 +2,7 @@ package raylras.zen.code.symbol;
 
 import org.antlr.v4.runtime.tree.ParseTree;
 import raylras.zen.code.CompilationUnit;
-import raylras.zen.code.type.resolve.DeclaratorResolver;
+import raylras.zen.code.resolve.DeclaratorResolver;
 import raylras.zen.code.scope.Scope;
 import raylras.zen.code.type.Type;
