+++ src/main/java/raylras/zen/util/MemberUtils.java
@@ -7,7 +7,7 @@ import raylras.zen.code.parser.ZenScriptParser;
 import raylras.zen.code.scope.Scope;
 import raylras.zen.code.symbol.*;
 import raylras.zen.code.type.*;
-import raylras.zen.code.type.resolve.ExpressionTypeResolver;
+import raylras.zen.code.resolve.ExpressionTypeResolver;
 import raylras.zen.code.CompilationEnvironment;
 import raylras.zen.service.MethodCallPriority;
 import raylras.zen.service.TypeService;
